Move config module to its own file
That way it can be reused and the structure is a bit more clear.
